refactor(gl): split GLCanvas platform bodies into per-platform TUs
