+++ src/CMakeLists.txt
@@ -84,11 +84,13 @@ if(CMAKE_SYSTEM_NAME STREQUAL "Emscripten")
   list(APPEND SOURCES
     platform/transport_factory_wasm.cpp
     platform/render_scheduler_wasm.cpp
+    ui/gl_canvas_wasm.cpp
   )
 else()
   list(APPEND SOURCES
     platform/transport_factory_native.cpp
     platform/render_scheduler_native.cpp
+    ui/gl_canvas_native.cpp
   )
 endif()
