+++ src/ui/gl_canvas.cpp
@@ -28,10 +28,6 @@
 #include <cmath>
 #include <iostream>
 
-#ifdef __EMSCRIPTEN__
-#include <rhi/qrhi.h>
-#endif
-
 #include "main_window/main_window.h"
 #include "ui/gl_text_renderer.h"
 #include "visualization/components/camera.h"
@@ -41,20 +37,10 @@
 namespace oid {
 
 GLCanvas::GLCanvas(QWidget* parent)
-#ifdef __EMSCRIPTEN__
-    : QRhiWidget{parent},
-#else
-    : QOpenGLWidget{parent},
-#endif
+    : GlWidgetBase{parent},
       text_renderer_{std::make_unique<GLTextRenderer>(*this)} {
     mouse_down_[0] = mouse_down_[1] = false;
-#ifdef __EMSCRIPTEN__
-    setApi(Api::OpenGL);
-    connect(this, &QRhiWidget::renderFailed, this, [this] {
-        initialized_ = false;
-        first_paint_completed_ = false;
-    });
-#endif
+    platform_ctor_init();
 }
 
 GLCanvas::~GLCanvas() = default;
@@ -161,189 +147,6 @@ void GLCanvas::destroy_icon_framebuffer() {
     }
     icon_framebuffer_ready_ = false;
 }
-
-#ifdef __EMSCRIPTEN__
-
-int GLCanvas::render_width() const {
-    if (const auto* tex = colorTexture(); tex != nullptr) {
-        return tex->pixelSize().width();
-    }
-    return static_cast<int>(std::lround(width() * devicePixelRatioF()));
-}
-
-int GLCanvas::render_height() const {
-    if (const auto* tex = colorTexture(); tex != nullptr) {
-        return tex->pixelSize().height();
-    }
-    return static_cast<int>(std::lround(height() * devicePixelRatioF()));
-}
-
-bool GLCanvas::init_gl_state() {
-    glEnable(GL_BLEND);
-    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-    glClearColor(0.1f, 0.1f, 0.1f, 1.0f);
-
-    if (!text_renderer_->initialize()) {
-        return false;
-    }
-
-    return true;
-}
-
-void GLCanvas::initialize(QRhiCommandBuffer* cb) {
-    static QRhi* last_rhi = nullptr;
-    if (last_rhi != rhi()) {
-        initialized_ = false;
-        first_paint_completed_ = false;
-        last_rhi = rhi();
-    }
-    if (initialized_) {
-        return;
-    }
-
-    const auto clear_color = QColor::fromRgbF(0.1f, 0.1f, 0.1f, 1.0f);
-    cb->beginPass(renderTarget(),
-                  clear_color,
-                  {1.0f, 0},
-                  nullptr,
-                  QRhiCommandBuffer::ExternalContent);
-    cb->beginExternal();
-    initializeOpenGLFunctions();
-
-    if (!init_gl_state()) {
-        cb->endExternal();
-        cb->endPass();
-        return;
-    }
-
-    init_icon_framebuffer();
-
-    cb->endExternal();
-    cb->endPass();
-    initialized_ = true;
-}
-
-void GLCanvas::render(QRhiCommandBuffer* cb) {
-    if (!initialized_) {
-        return;
-    }
-
-    const auto clear_color = QColor::fromRgbF(0.1f, 0.1f, 0.1f, 1.0f);
-    cb->beginPass(renderTarget(),
-                  clear_color,
-                  {1.0f, 0},
-                  nullptr,
-                  QRhiCommandBuffer::ExternalContent);
-    cb->beginExternal();
-
-    drain_gl_queue();
-
-    const auto w = render_width();
-    const auto h = render_height();
-    glViewport(0, 0, w, h);
-    if (w != last_render_width_ || h != last_render_height_) {
-        last_render_width_ = w;
-        last_render_height_ = h;
-        main_window().resize_callback(w, h);
-    }
-    main_window().prepare_gl_draw();
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-    main_window().draw();
-    drain_icon_gl_queue();
-
-    cb->endExternal();
-    cb->endPass();
-
-    first_paint_completed_ = true;
-}
-
-void GLCanvas::releaseResources() {
-    initialized_ = false;
-    first_paint_completed_ = false;
-    last_render_width_ = 0;
-    last_render_height_ = 0;
-    destroy_icon_framebuffer();
-}
-
-void GLCanvas::resizeEvent(QResizeEvent* e) {
-    QRhiWidget::resizeEvent(e);
-    update();
-}
-
-void GLCanvas::schedule_gl(std::function<void()> task) {
-    gl_queue_.push_back(std::move(task));
-    update();
-}
-
-void GLCanvas::schedule_icon_gl(std::function<void()> task) {
-    icon_gl_queue_.push_back(std::move(task));
-    update();
-}
-
-void GLCanvas::drain_gl_queue() {
-    while (!gl_queue_.empty()) {
-        auto task = std::move(gl_queue_.front());
-        gl_queue_.pop_front();
-        task();
-    }
-}
-
-void GLCanvas::drain_icon_gl_queue() {
-    while (!icon_gl_queue_.empty()) {
-        auto task = std::move(icon_gl_queue_.front());
-        icon_gl_queue_.pop_front();
-        task();
-    }
-}
-
-#else
-
-int GLCanvas::render_width() const {
-    return width();
-}
-
-int GLCanvas::render_height() const {
-    return height();
-}
-
-void GLCanvas::initializeGL() {
-    this->makeCurrent();
-    if (const auto context = this->context();
-        context == nullptr || !context->isValid()) [[unlikely]] {
-        std::cerr << "[Error] OpenGL context is not valid. OpenGL "
-                     "initialization cannot proceed."
-                  << std::endl;
-        initialized_ = false;
-        return;
-    }
-    initializeOpenGLFunctions();
-
-    glEnable(GL_BLEND);
-    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-    glClearColor(0.1f, 0.1f, 0.1f, 1.0f);
-
-    init_icon_framebuffer();
-
-    // Initialize text renderer
-    if (!text_renderer_->initialize()) {
-        initialized_ = false;
-        return;
-    }
-
-    initialized_ = true;
-}
-
-void GLCanvas::paintGL() {
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-    main_window().draw();
-}
-
-void GLCanvas::resizeGL(const int w, const int h) {
-    glViewport(0, 0, w, h);
-    main_window().resize_callback(w, h);
-}
-
-#endif
 
 void GLCanvas::wheelEvent(QWheelEvent* ev) {
     main_window().scroll_callback(static_cast<float>(ev->angleDelta().y()) /
