+++ src/ui/main_window/main_window.h
@@ -116,9 +116,7 @@ class MainWindow final : public QMainWindow {
 
     void draw() const;
 
-#ifdef __EMSCRIPTEN__
     void prepare_gl_draw() const;
-#endif
 
     [[nodiscard]] std::shared_ptr<GLCanvas> gl_canvas() const;
